+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * The minimum support version of React is 16.8.0
 * Killed the possibility to directly pass store names to `observer`. Always use `inject` instead. (This was deprecate for a long time already). `observer(["a", "b"], component)` should now be written as `inject("a", "b")(component)`.
 * `observer` components no longer automatically recover from errors (to prevent potential memory leaks). Instead, this is the responsibility of error boundaries.
+* `inject` now supports ref forwarding. As such, the `.wrappedInstance` property has been removed since refs can be used instead.
 
 **Improvements**